Mostly fleshing out daemon more
Ended up reworking the way daemons are loaded so that it operates as a singleton object that masks as a function. The first time you require it takes config params, and then sets up / caches and returns that for subsequent invocations. The submission of a job has a lot of moving parts (and still isn't done), I'm trying to be very careful about where code goes.

The current idea is that subclasses of daemon implement (at least) two functions; on_tick and submit_job. on_tick needs to sync the MongoDB state with the underlying job queue. submit_job takes a valid Run document from MongoDB, and pushes it into the job queue.

For the PBS server, this is going to involve the generation of a PBS script, and then pushing the job into PBS and saving the corresponding PBS ID on the Run document. The webserver subclasses (ie. rnamutants/index.js) must provide a function to pbs_command that takes a Run document and returns the corresponding command line call to make. The PBS subclass of the daemon object handles the generation of the surrounding scaffold for the command (PBS directives, etc.) and the creation of the workspace where the job is run.

It also *will* handle the actual submission and updating in MongoDB, but I'm not there yet.

Signed-off-by: Evan Senter <evansenter@gmail.com>